# deal/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
    title: Deal API
    version: 1.0.0
paths:
    /deal/statement:
        post:
            tags:
                - deal-controller
            operationId: statementProcessing
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoanStatementRequestDto"
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        "*/*":
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/LoanOfferDto"
    /deal/offer/select:
        post:
            tags:
                - deal-controller
            operationId: selectOffer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoanOfferDto"
                required: true
            responses:
                "200":
                    description: OK
    "/deal/calculate/{statementId}":
        post:
            tags:
                - deal-controller
            operationId: registrationComplete
            parameters:
                - name: statementId
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/FinishRegistrationRequestDto"
                required: true
            responses:
                "200":
                    description: OK
    "/deal/document/{statementId}/send":
        post:
            tags:
                - deal-controller
            operationId: sendDocuments
            parameters:
                - name: statementId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
    "/deal/document/{statementId}/sign":
        post:
            tags:
                - deal-controller
            operationId: signDocuments
            parameters:
                - name: statementId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
    "/deal/document/{statementId}/code":
        post:
            tags:
                - deal-controller
            operationId: codeDocuments
            parameters:
                - name: statementId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
    "/deal/admin/statement/{statementId}/status":
        put:
            tags:
                - deal-controller
            operationId: updateStatementStatus
            parameters:
                - name: statementId
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                description: "Application statuses: PREAPPROVAL, APPROVED, CC_DENIED, CC_APPROVED, PREPARE_DOCUMENTS, DOCUMENT_CREATED, CLIENT_DENIED, DOCUMENT_SIGNED, CREDIT_ISSUED"
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ApplicationStatus"
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/StatementDto"
    "/deal/admin/statement/{statementId}":
        get:
            tags:
                - deal-controller
            operationId: getStatement
            parameters:
                - name: statementId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/StatementDto"
components:
    schemas:
        LoanStatementRequestDto:
            type: object
            properties:
                amount:
                    type: number
                    example: 150000
                term:
                    type: integer
                    format: int32
                    example: 12
                firstName:
                    type: string
                    example: Иван
                lastName:
                    type: string
                    example: Иванов
                middleName:
                    type: string
                    example: Иванович
                email:
                    type: string
                    example: 3J7wI@example.com
                birthDate:
                    type: string
                    format: date
                    example: 2000-01-01
                passportSeries:
                    type: string
                    example: 1234
                passportNumber:
                    type: string
                    example: 123456

        LoanOfferDto:
            type: object
            properties:
                statementId:
                    type: string
                    format: uuid
                requestedAmount:
                    type: number
                totalAmount:
                    type: number
                term:
                    type: integer
                    format: int32
                monthlyPayment:
                    type: number
                rate:
                    type: number
                isInsuranceEnabled:
                    type: boolean
                isSalaryClient:
                    type: boolean

        EmploymentDto:
            type: object
            properties:
                employmentStatus:
                    $ref: "#/components/schemas/EmploymentStatus"
                employerINN:
                    type: string
                    example: 123456789021
                salary:
                    type: number
                    example: 43000
                position:
                    $ref: "#/components/schemas/Position"
                workExperienceTotal:
                    type: integer
                    format: int32
                    example: 24
                workExperienceCurrent:
                    type: integer
                    format: int32
                    example: 12

        ScoringDataDto:
            type: object
            properties:
                amount:
                    type: number
                term:
                    type: integer
                    format: int32
                firstName:
                    type: string
                lastName:
                    type: string
                middleName:
                    type: string
                gender:
                    $ref: "#/components/schemas/Gender"
                birthdate:
                    type: string
                    format: date
                passportSeries:
                    type: string
                passportNumber:
                    type: string
                passportIssueDate:
                    type: string
                    format: date
                passportIssueBranch:
                    type: string
                maritalStatus:
                    $ref: "#/components/schemas/MaritalStatus"
                dependentAmount:
                    type: integer
                    format: int32
                employment:
                    $ref: "#/components/schemas/EmploymentDto"
                accountNumber:
                    type: string
                isInsuranceEnabled:
                    type: boolean
                isSalaryClient:
                    type: boolean

        PaymentScheduleElementDto:
            type: object
            properties:
                number:
                    type: integer
                    format: int32
                date:
                    type: string
                    format: date
                totalPayment:
                    type: number
                interestPayment:
                    type: number
                debtPayment:
                    type: number
                remainingDebt:
                    type: number

        StatementDto:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                status:
                    $ref: "#/components/schemas/ApplicationStatus"
                creationDate:
                    type: string
                    format: date
                appliedOffer:
                    $ref: "#/components/schemas/LoanOfferDto"
                signDate:
                    type: string
                    format: date
                statusHistory:
                    type: array
                    items:
                        $ref: "#/components/schemas/StatementStatusHistoryDto"

        StatementStatusHistoryDto:
            type: object
            properties:
                status:
                    $ref: "#/components/schemas/ApplicationStatus"
                time:
                    type: string
                    format: date-time
                changeType:
                    $ref: "#/components/schemas/ChangeType"

        FinishRegistrationRequestDto:
            type: object
            properties:
                gender:
                    $ref: "#/components/schemas/Gender"
                maritalStatus:
                    $ref: "#/components/schemas/MaritalStatus"
                dependentAmount:
                    type: integer
                    format: int32
                    example: 1
                passportIssueDate:
                    type: string
                    format: date
                    example: 2010-01-01
                passportIssueBranch:
                    type: string
                    example: УФМС г. Москвы
                employment:
                    $ref: "#/components/schemas/EmploymentDto"
                accountNumber:
                    type: string
                    example: 12345678901234567890

        CreditDto:
            type: object
            properties:
                amount:
                    type: number
                term:
                    type: integer
                    format: int32
                monthlyPayment:
                    type: number
                rate:
                    type: number
                psk:
                    type: number
                isInsuranceEnabled:
                    type: boolean
                isSalaryClient:
                    type: boolean
                paymentSchedule:
                    type: array
                    items:
                        $ref: "#/components/schemas/PaymentScheduleElementDto"

        EmailMessage:
            type: object
            properties:
                address:
                    type: string
                theme:
                    $ref: "#/components/schemas/EmailTheme"
                statementId:
                    type: string
                text:
                    type: string

        Gender:
            type: string
            enum: [MALE, FEMALE]
        MaritalStatus:
            type: string
            enum: [SINGLE, MARRIED, DIVORCED, WIDOWED]
        EmploymentStatus:
            type: string
            enum: [EMPLOYED, UNEMPLOYED, SELF_EMPLOYED, BUSINESS_OWNER]
        Position:
            type: string
            enum: [MIDDLE_MANAGER, TOP_MANAGER]
        CreditStatus:
            type: string
            enum: [CALCULATED, ISSUED]
        ApplicationStatus:
            type: string
            enum:
                [
                    PREAPPROVAL,
                    APPROVED,
                    CC_DENIED,
                    CC_APPROVED,
                    PREPARE_DOCUMENTS,
                    DOCUMENT_CREATED,
                    CLIENT_DENIED,
                    DOCUMENT_SIGNED,
                    CREDIT_ISSUED,
                ]
        ChangeType:
            type: string
            enum: [AUTOMATIC, MANUAL]
        EmailTheme:
            type: string
            enum:
                [
                    FINISH_REGISTRATION,
                    REGISTRATION_DOCUMENTS,
                    DOCUMENT_CREATED,
                    SIGN_DOCUMENTS,
                    CREDIT_ISSUED,
                ]